Adding moving events and implementing a pseudo-controller in Actor class (Need to create a controller component)

+++ include/GameInfo.h
@@ -12,7 +12,7 @@ struct GameInfo
 	bool m_run = true;
 	uint8_t m_gameState = STATE_PLAY;
 
-	float m_lastTime = 0;
-	float m_thisTime = 0;
-	float m_deltaTime = 0;
+	double m_lastTime = 0;
+	double m_thisTime = 0;
+	double m_deltaTime = 0;
 };
